Change version_scheme to 1 for vasmm68k
The first commits of vasmm68k accidentally werw without any explicit version. The automatic version detection failed and thought vasm1_8f meant version '8f'. Now that we are providing an explicit version, we also need to bump the version scheme so it doesn't seem like 8f -> 1.8g is a downgrade.